+++ test/test_utilities.cpp
@@ -52,7 +52,7 @@ protected:
 	static void SetUpTestCase() {
 		N = 10000;
 		d = 128;
-		int i, j;
+		size_t i, j;
 
 		// For generating random numbers
 		random_device rd;
@@ -87,13 +87,13 @@ protected:
 public:
 	// Some expensive resource shared by all tests.
 	static float ** data;
-	static int N;
-	static int d;
+	static size_t N;
+	static size_t d;
 };
 
 float ** UtilTest::data;
-int UtilTest::N;
-int UtilTest::d;
+size_t UtilTest::N;
+size_t UtilTest::d;
 
 // Let's start with some testcases
 TEST_F(UtilTest, test1) {
@@ -115,9 +115,9 @@ TEST_F(UtilTest, test4) {
 }
 
 TEST_F(UtilTest, test5) {
-	int index[] = {0,1,2,3,4};
-	int * index2 = (int *)malloc(5 * sizeof(int));
-	for(int i = 0; i < 5; i++)
+	size_t index[] = {0,1,2,3,4};
+	size_t * index2 = (size_t *)malloc(5 * sizeof(size_t));
+	for(size_t i = 0; i < 5; i++)
 		index2[i] = i;
 
 	float * tmp = (float *)calloc(d,sizeof(float));
@@ -154,7 +154,7 @@ TEST_F(UtilTest, test9) {
 	float * t;
 	EXPECT_TRUE(init_array<float>(t,d) && t != nullptr);
 	EXPECT_TRUE(copy_array<float>(data[0],t,d));
-	for(int i = 0; i < d; i++) {
+	for(size_t i = 0; i < d; i++) {
 		EXPECT_TRUE(data[0][i] == t[i]);
 	}
 }
@@ -168,8 +168,8 @@ TEST_F(UtilTest, test11) {
 	float ** t;
 	EXPECT_TRUE(init_array_2<float>(t,N,d) && t != nullptr);
 	EXPECT_TRUE(copy_array_2<float>(data,t,N,d));
-	for(int i = 0; i < N; i++) {
-		for(int j = 0; j < d; j++) {
+	for(size_t i = 0; i < N; i++) {
+		for(size_t j = 0; j < d; j++) {
 			EXPECT_TRUE(data[i][j] == t[i][j]);
 		}
 	}
